+++ src/careamics_restoration/dataset/__init__.py
@@ -1,5 +1,5 @@
-from .dataloader_factory import create_dataset
-from .dataloader_utils import (
+from .dataset_factory import create_dataset
+from .dataset_utils import (
     compute_crop_and_stitch_coords_1d,
     compute_overlap,
     compute_patch_steps,
